Some new matrix-handling features.
> Matrix.Area(x1,y1,x2,y2:Int32):<Matrix>;
> Matrix.Cols(FromCol,ToCol:Int32): <Matrix>;
> Matrix.Rows(FromRow,ToRow:Int32): <Matrix>;

Also added a better way of raising errors in the plugin. Basicly trow a
function-call in a `try .. except` clause and in the `except`-part you
can call `WriteLn( se.GetException() );` to get the error raised in the
plugin. Some functions already support this.. Tho does only contain the
exceptions that is raised by "hand", not the FPC internal exceptions.